+++ SEAL/.idea/misc.xml
@@ -12,10 +12,17 @@
         <entry key="..\:/Android/Project/SE_AL-front/SE_AL/SEAL/app/src/main/res/layout/fragment_setting.xml" value="0.26874366767983787" />
         <entry key="..\:/Android/Project/SE_AL-front/SE_AL/SEAL/app/src/main/res/mipmap-anydpi-v26/ic_launcher.xml" value="0.5697916666666667" />
         <entry key="..\:/Users/hyxxsj/Desktop/SE_AL/SEAL/app/src/main/res/layout/activity_main.xml" value="0.25364583333333335" />
-        <entry key="..\:/Users/hyxxsj/Desktop/SE_AL/SEAL/app/src/main/res/layout/fragment_calendar_item.xml" value="0.4" />
+        <entry key="..\:/Users/hyxxsj/Desktop/SE_AL/SEAL/app/src/main/res/layout/fragment_calendar_item.xml" value="0.264" />
+        <entry key="..\:/Users/hyxxsj/Desktop/SE_AL/SEAL/app/src/main/res/layout/fragment_class.xml" value="0.25364583333333335" />
+        <entry key="..\:/Users/hyxxsj/Desktop/SE_AL/SEAL/app/src/main/res/layout/fragment_class_alarm_setting_base.xml" value="0.264" />
+        <entry key="..\:/Users/hyxxsj/Desktop/SE_AL/SEAL/app/src/main/res/layout/fragment_class_base.xml" value="0.2" />
+        <entry key="..\:/Users/hyxxsj/Desktop/SE_AL/SEAL/app/src/main/res/layout/fragment_class_list.xml" value="0.25" />
+        <entry key="..\:/Users/hyxxsj/Desktop/SE_AL/SEAL/app/src/main/res/layout/fragment_class_ui_base.xml" value="0.264" />
         <entry key="..\:/Users/hyxxsj/Desktop/SE_AL/SEAL/app/src/main/res/layout/fragment_dashboard.xml" value="0.1296875" />
+        <entry key="..\:/Users/hyxxsj/Desktop/SE_AL/SEAL/app/src/main/res/layout/fragment_home.xml" value="0.25364583333333335" />
         <entry key="..\:/Users/hyxxsj/Desktop/SE_AL/SEAL/app/src/main/res/layout/fragment_login.xml" value="0.264" />
         <entry key="..\:/Users/hyxxsj/Desktop/SE_AL/SEAL/app/src/main/res/layout/fragment_main.xml" value="0.2" />
+        <entry key="..\:/Users/hyxxsj/Desktop/SE_AL/SEAL/app/src/main/res/layout/fragment_setting.xml" value="0.1" />
       </map>
     </option>
   </component>
