+++ SEAL/app/src/main/java/com/example/seal/ui/main/ClassAlarmSettingBaseFragment.kt
@@ -1,14 +1,13 @@
 package com.example.seal.ui.main
 
 import android.os.Bundle
-import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.fragment.app.Fragment
 import com.example.seal.R
 
-
-class SettingFragment : Fragment() {
+class ClassAlarmSettingBaseFragment : Fragment() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -20,10 +19,12 @@ class SettingFragment : Fragment() {
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_setting, container, false)
+        return inflater.inflate(R.layout.fragment_class_alarm_setting_base, container, false)
     }
 
     companion object {
 
     }
+
+
 }
