+++ SEAL/app/src/main/java/com/example/seal/ui/main/ClassListFragment.kt
@@ -16,10 +16,10 @@ private const val ARG_PARAM2 = "param2"
 
 /**
  * A simple [Fragment] subclass.
- * Use the [ClassFragment.newInstance] factory method to
+ * Use the [ClassListFragment.newInstance] factory method to
  * create an instance of this fragment.
  */
-class ClassFragment : Fragment() {
+class ClassListFragment : Fragment() {
     // TODO: Rename and change types of parameters
     private var param1: String? = null
     private var param2: String? = null
@@ -37,7 +37,7 @@ class ClassFragment : Fragment() {
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_class, container, false)
+        return inflater.inflate(R.layout.fragment_class_list, container, false)
 
         val fab: FloatingActionButton
 
@@ -54,12 +54,12 @@ class ClassFragment : Fragment() {
          *
          * @param param1 Parameter 1.
          * @param param2 Parameter 2.
-         * @return A new instance of fragment ClassFragment.
+         * @return A new instance of fragment ClassListFragment.
          */
         // TODO: Rename and change types and number of parameters
         @JvmStatic
         fun newInstance(param1: String, param2: String) =
-            ClassFragment().apply {
+            ClassListFragment().apply {
                 arguments = Bundle().apply {
                     putString(ARG_PARAM1, param1)
                     putString(ARG_PARAM2, param2)
