+++ SEAL/app/src/main/java/com/example/seal/ui/main/DashboardFragment.kt
@@ -1,0 +1,6 @@
+package com.example.seal.ui.main
+
+import androidx.fragment.app.Fragment
+
+class DashboardFragment : Fragment() {
+}
