+++ SEAL/app/src/main/res/layout/fragment_class_list.xml
@@ -1,5 +1,4 @@
 <?xml version="1.0" encoding="utf-8"?>
-
 <androidx.constraintlayout.widget.ConstraintLayout
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
@@ -15,26 +14,32 @@
         app:layout_constraintTop_toTopOf="parent">
 
         <Button
-            android:id="@+id/alarm_time"
+            android:id="@+id/class1"
             android:layout_width="300dp"
             android:layout_height="50dp"
             android:layout_marginBottom="20dp"
-            android:text="@string/alarm_time_kr" />
+            android:text="@string/class1" />
 
         <Button
-            android:id="@+id/export_calendar"
+            android:id="@+id/class2"
             android:layout_width="300dp"
             android:layout_height="50dp"
             android:layout_marginBottom="20dp"
-            android:text="@string/export_calendar" />
+            android:text="@string/class2" />
 
         <Button
-            android:id="@+id/personal_info_modify"
+            android:id="@+id/class3"
             android:layout_width="300dp"
             android:layout_height="50dp"
             android:layout_marginBottom="20dp"
-            android:text="@string/personal_info_modify" />
+            android:text="@string/class3" />
 
+        <Button
+            android:id="@+id/class4"
+            android:layout_width="300dp"
+            android:layout_height="50dp"
+            android:layout_marginBottom="20dp"
+            android:text="@string/class4" />
 
     </LinearLayout>
 </androidx.constraintlayout.widget.ConstraintLayout>
