--- SEAL/app/src/main/res/layout/fragment_main.xml
@@ -1,53 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<layout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools">
-
-    <androidx.constraintlayout.widget.ConstraintLayout
-        android:id="@+id/constraintLayout"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent" >
-
-        <!-- 화면전환 -->
-        <Button
-            android:id="@+id/screen_change"
-            android:layout_width="80dp"
-            android:layout_height="40dp"
-            android:layout_marginEnd="50dp"
-            app:layout_constraintBottom_toTopOf="@+id/calendar"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintTop_toTopOf="parent"
-            android:text="@string/screen_change"
-            android:textSize="12sp" />
-
-        <!-- 캘린더 -->
-        <CalendarView
-            android:id="@+id/calendar"
-            android:layout_width="300dp"
-            android:layout_height="300dp"
-            android:layout_marginTop="100dp"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent" />
-
-
-        <!-- 날짜 클릭 -->
-        <androidx.recyclerview.widget.RecyclerView
-            android:id="@+id/calendar_item"
-            android:layout_width="300dp"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="20dp"
-            android:paddingBottom="?attr/actionBarSize"
-            android:scrollbars="vertical"
-            app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@id/calendar"
-            tools:listitem="@layout/fragment_calendar_item" />
-
-    </androidx.constraintlayout.widget.ConstraintLayout>
-
-</layout>
-
-
