+++ SEAL/app/src/main/res/values/strings.xml
@@ -3,16 +3,47 @@
     <string name="tab_text_1">강의 목록</string>
     <string name="tab_text_2">home</string>
     <string name="tab_text_3">설정</string>
- 
+
 
     <string name="calendar_item_title">calendar item title</string>
     <string name="calendar_item_context">calendar item context</string>
     <string name="screen_change">화면전환</string>
     <string name="add_item">add</string>
-<!--    <string name=""></string>-->
 
- 
+
+
     <!-- TODO: Remove or change this placeholder text -->
     <string name="hello_blank_fragment">Hello blank fragment</string>
- 
+
+
+    <!--  class list fragment  -->
+    <string name="class1">class 1</string>
+    <string name="class2">class 2</string>
+    <string name="class3">class 3</string>
+    <string name="class4">class 4</string>
+
+
+    <!--  class base fragment  -->
+    <string name="add_item_kr">직접입력</string>
+    <string name="setting_kr">알림설정</string>
+    <string name="exam">시험</string>
+    <string name="homework">과제</string>
+    <string name="online_lecture">강의</string>
+    <string name="notice">공지사항</string>
+
+
+    <!--  setting fragment  -->
+    <string name="alarm_time_kr">알람 시간 설정</string>
+    <string name="export_calendar">캘린더 내보내기</string>
+    <string name="personal_info_modify">개인정보 수정</string>
+
+    <!--  class alarm setting base fragment  -->
+    <string name="class1_alarm">class1 - 알림 설정</string>
+    <string name="switch_notice_kr">공지사항 알림</string>
+    <string name="switch_all_kr">전체 알림</string>
+    <string name="switch_online_lecture">강의 알림</string>
+    <string name="switch_homework">과제 알림</string>
+    <string name="switch_exam">시험 알림</string>
+
+
 </resources>
